requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
langchain==0.3.27
langchain-openai==0.3.33
langchain-core==0.3.76
langgraph==0.6.8
pydantic==2.11.9
pydantic-core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
httpx==0.28.1
aiofiles==25.1.0